auth/requirements.txt
flask==2.2.2
gevent==21.12.0
gunicorn==20.1.0
sqlalchemy==1.4.40
Flask-SQLAlchemy==2.5.1
sqlalchemy-utils==0.38.3
flask-jwt-extended==4.4.4
passlib==1.7.4
redis==4.3.4
flask-redis==0.4.0
pydantic==1.10.0
psycopg2==2.9.3
apiflask[yaml]==1.1.2
alembic==1.8.1
opentelemetry-api==1.12.0
opentelemetry-sdk==1.12.0
opentelemetry-instrumentation-flask==0.33b0
opentelemetry-exporter-jaeger==1.12.0
Flask-Limiter==2.6.2
requests==2.28.1
oauthlib==3.2.1
pyOpenSSL==22.0.0
user-agents==2.2.0
